Removed old `compile` pass.
Compilation now has an intermediate pass that performs all type
conversions and monomorphization but keeps higher-level function calls,
returns, and SSA for local variables.
