+++ src/compile.rs
@@ -1,15 +1,12 @@
-use hashbrown::{hash_map::Entry, HashMap, HashSet};
-use smallvec::smallvec;
-
 use crate::builtins::{self, Variable};
-use crate::bytecode::{self, Accum, Instr, Interp};
+use crate::bytecode;
 use crate::cfg::{self, is_unused, Function, Ident, PrimExpr, PrimStmt, PrimVal, ProgramContext};
-use crate::common::{NodeIx, NumTy, Result};
-use crate::types::{self, SmallVec};
+use crate::common::{Either, Graph, NodeIx, NumTy, Result};
+use crate::smallvec::{self, smallvec};
+use crate::types;
 
-// TODO: implement basic "optimizations"
-//    * avoid excessive moves (unless those come from the cfg?)
-//    * don't emit jumps that just point to the next instruction.
+use hashbrown::{hash_map::Entry, HashMap};
+use std::mem;
 
 pub(crate) const UNUSED: u32 = u32::max_value();
 
@@ -35,7 +32,7 @@ impl Default for Ty {
 }
 
 impl Ty {
-    pub(crate) fn of_var(v: Variable) -> Ty {
+    fn of_var(v: Variable) -> Ty {
         use Variable::*;
         match v {
             ARGC | FS | OFS | RS | FILENAME => Ty::Str,
@@ -44,7 +41,7 @@ impl Ty {
         }
     }
 
-    pub(crate) fn iter(self) -> Result<Ty> {
+    fn iter(self) -> Result<Ty> {
         use Ty::*;
         match self {
             IterInt => Ok(Int),
@@ -54,7 +51,7 @@ impl Ty {
         }
     }
 
-    pub(crate) fn key(self) -> Result<Ty> {
+    fn key(self) -> Result<Ty> {
         use Ty::*;
         match self {
             MapIntInt | MapIntFloat | MapIntStr => Ok(Int),
@@ -64,7 +61,8 @@ impl Ty {
             }
         }
     }
-    pub(crate) fn val(self) -> Result<Ty> {
+
+    fn val(self) -> Result<Ty> {
         use Ty::*;
         match self {
             MapStrInt | MapIntInt => Ok(Int),
@@ -79,19 +77,39 @@ impl Ty {
 
 pub(crate) const NUM_TYPES: usize = Ty::IterStr as usize + 1;
 
-// This is a macro to defeat the borrow checker when used inside methods for `Generator`.
-macro_rules! reg_of_ty {
-    ($slf:expr, $ty:expr) => {{
-        $slf.reg_stats.push($ty, RegStatus::Local)
-    }};
-    ($slf:expr, $ty:expr, $status:expr) => {{
-        $slf.reg_stats.push($ty, $status)
-    }};
+pub(crate) fn bytecode<'a>(
+    ctx: &cfg::ProgramContext<'a, &'a str>,
+    // default to std::io::stdin()
+    reader: impl std::io::Read + 'static,
+    // default to std::io::BufWriter::new(std::io::stdout())
+    writer: impl std::io::Write + 'static,
+) -> Result<bytecode::Interp<'a>> {
+    Typer::init_from_ctx(ctx)?.to_interp(reader, writer)
+}
+type SmallVec<T> = smallvec::SmallVec<[T; 2]>;
+
+#[derive(Debug)]
+enum HighLevel {
+    // TODO we may not strictly need Call's dst_ty and Ret's Ty field. Other information may have
+    // it available.
+    Call {
+        func_id: NumTy, /* monomorphized function id */
+        dst_reg: NumTy,
+        dst_ty: Ty,
+        args: SmallVec<(NumTy, Ty)>,
+    },
+    Ret(NumTy, Ty),
+    Phi(NumTy, Ty, SmallVec<(NodeIx /*pred*/, NumTy /*register*/)>),
+}
+
+#[derive(Default)]
+struct Registers {
+    stats: RegStatuses,
+    globals: HashMap<Ident, (u32, Ty)>,
 }
 
 #[derive(Debug, Copy, Clone)]
 enum RegStatus {
-    // Arg?
     Local,
     Global,
     Ret,
@@ -101,7 +119,10 @@ enum RegStatus {
 struct RegStatuses([Vec<RegStatus>; NUM_TYPES]);
 
 impl RegStatuses {
-    fn push(&mut self, ty: Ty, status: RegStatus) -> NumTy {
+    fn reg_of_ty(&mut self, ty: Ty) -> NumTy {
+        self.new_reg(ty, RegStatus::Local)
+    }
+    fn new_reg(&mut self, ty: Ty, status: RegStatus) -> NumTy {
         let v = &mut self.0[ty as usize];
         let res = v.len();
         v.push(status);
@@ -117,32 +138,13 @@ impl RegStatuses {
     }
 }
 
-#[derive(Debug)]
-struct FuncInfo {
-    ret_reg: NumTy,
-    ret_ty: Ty,
-    // The following are only used for codegen purposes; they do not impact the generated bytecode.
-    args: SmallVec<Ty>, // arity == args.len()
-}
-
-macro_rules! dbg_reg_valid {
-    ($gen:expr, $reg:expr, $ty:expr) => {
-        #[cfg(debug_assertions)]
-        {
-            let reg = $reg;
-            let ty = $ty;
-            let count = $gen.regs.reg_stats.count(ty);
-            if reg != UNUSED {
-                assert!(reg < count, "[{} : {:?}] vs. {}", reg, ty, count);
-            }
-        }
-    };
-}
+type LL<'a> = bytecode::Instr<'a>;
+type Instr<'a> = Either<LL<'a>, HighLevel>;
+type CFG<'a> = Graph<Vec<Instr<'a>>, Option<NumTy /* Int register */>>;
 
 #[derive(Default)]
-struct ProgramGenerator<'a> {
+struct Typer<'a> {
     regs: Registers,
-
     id_map: HashMap<
         // TODO: make newtypes for these different Ids?
         (
@@ -166,27 +168,265 @@ struct ProgramGenerator<'a> {
     main_offset: usize,
 }
 
-impl<'a> ProgramGenerator<'a> {
-    fn init_from_ctx(pc: &ProgramContext<'a, &'a str>) -> Result<ProgramGenerator<'a>> {
-        // Type-check the code, then initialize a ProgramGenerator, assigning registers to local
+#[derive(Debug)]
+struct FuncInfo {
+    ret_ty: Ty,
+    // For bytecode, we pop into each of these registers at the specified type.
+    arg_tys: SmallVec<Ty>,
+}
+
+#[derive(Default)]
+struct Frame<'a> {
+    src_function: NumTy,
+    cur_ident: NumTy,
+    locals: HashMap<Ident, (u32, Ty)>,
+    arg_regs: SmallVec<NumTy>,
+    cfg: CFG<'a>,
+}
+
+struct View<'a, 'b> {
+    frame: &'b mut Frame<'a>,
+    regs: &'b mut Registers,
+    id_map: &'b HashMap<(NumTy, SmallVec<Ty>), NumTy>,
+    arity: &'b HashMap<NumTy, NumTy>,
+    func_info: &'b Vec<FuncInfo>,
+    // The current basic block being filled; It'll be swaped into `frame.cfg` as we translate a
+    // given function cfg.
+    stream: &'b mut Vec<Instr<'a>>,
+}
+
+fn pop_var<'a>(instrs: &mut Vec<LL<'a>>, reg: NumTy, ty: Ty) -> Result<()> {
+    use Ty::*;
+    instrs.push(match ty {
+        Int => LL::PopInt(reg.into()),
+        Float => LL::PopFloat(reg.into()),
+        Str => LL::PopStr(reg.into()),
+        MapIntInt => LL::PopIntInt(reg.into()),
+        MapIntFloat => LL::PopIntFloat(reg.into()),
+        MapIntStr => LL::PopIntStr(reg.into()),
+        MapStrInt => LL::PopStrInt(reg.into()),
+        MapStrFloat => LL::PopStrFloat(reg.into()),
+        MapStrStr => LL::PopStrStr(reg.into()),
+        IterInt | IterStr => return err!("invalid argument type: {:?}", ty),
+    });
+    Ok(())
+}
+
+fn push_var<'a>(instrs: &mut Vec<LL<'a>>, reg: NumTy, ty: Ty) -> Result<()> {
+    use Ty::*;
+    instrs.push(match ty {
+        Int => LL::PushInt(reg.into()),
+        Float => LL::PushFloat(reg.into()),
+        Str => LL::PushStr(reg.into()),
+        MapIntInt => LL::PushIntInt(reg.into()),
+        MapIntFloat => LL::PushIntFloat(reg.into()),
+        MapIntStr => LL::PushIntStr(reg.into()),
+        MapStrInt => LL::PushStrInt(reg.into()),
+        MapStrFloat => LL::PushStrFloat(reg.into()),
+        MapStrStr => LL::PushStrStr(reg.into()),
+        IterInt | IterStr => return err!("invalid argument type: {:?}", ty),
+    });
+    Ok(())
+}
+
+fn mov<'a>(dst_reg: u32, src_reg: u32, ty: Ty) -> Result<Option<LL<'a>>> {
+    use Ty::*;
+    if dst_reg == UNUSED || src_reg == UNUSED {
+        return Ok(None);
+    }
+
+    let res = match ty {
+        Int => LL::MovInt(dst_reg.into(), src_reg.into()),
+        Float => LL::MovFloat(dst_reg.into(), src_reg.into()),
+        Str => LL::MovStr(dst_reg.into(), src_reg.into()),
+
+        MapIntInt => LL::MovMapIntInt(dst_reg.into(), src_reg.into()),
+        MapIntFloat => LL::MovMapIntFloat(dst_reg.into(), src_reg.into()),
+        MapIntStr => LL::MovMapIntStr(dst_reg.into(), src_reg.into()),
+
+        MapStrInt => LL::MovMapStrInt(dst_reg.into(), src_reg.into()),
+        MapStrFloat => LL::MovMapStrFloat(dst_reg.into(), src_reg.into()),
+        MapStrStr => LL::MovMapStrStr(dst_reg.into(), src_reg.into()),
+
+        IterInt | IterStr => return err!("attempt to move values of type {:?}", ty),
+    };
+    Ok(Some(res))
+}
+
+impl<'a> Typer<'a> {
+    fn to_interp(
+        &mut self,
+        reader: impl std::io::Read + 'static,
+        writer: impl std::io::Write + 'static,
+    ) -> Result<bytecode::Interp<'a>> {
+        let instrs = self.to_bytecode()?;
+        Ok(bytecode::Interp::new(
+            instrs,
+            self.main_offset,
+            |ty| self.regs.stats.count(ty) as usize,
+            reader,
+            writer,
+        ))
+    }
+    fn to_bytecode(&mut self) -> Result<Vec<Vec<LL<'a>>>> {
+        let mut res = vec![vec![]; self.frames.len()];
+        let ret_regs: Vec<_> = (0..self.frames.len())
+            .map(|i| {
+                let ret_ty = self.func_info[i].ret_ty;
+                self.regs.stats.new_reg(ret_ty, RegStatus::Ret)
+            })
+            .collect();
+        let mut bb_map: Vec<usize> = Vec::new();
+        let mut jmps: Vec<usize> = Vec::new();
+        // If we wanted to, we could colocate locals and args, but absent a serious performance
+        // issue this seems cleaner.
+        let mut args: Vec<(NumTy, Ty)> = Vec::new();
+        let mut locals: Vec<(NumTy, Ty)> = Vec::new();
+        for (i, frame) in self.frames.iter().enumerate() {
+            let instrs = &mut res[i];
+            bb_map.clear();
+            bb_map.reserve(frame.cfg.node_count());
+            jmps.clear();
+
+            // Start by popping any args off of the stack.
+            args.extend(
+                frame
+                    .arg_regs
+                    .iter()
+                    .cloned()
+                    .zip(self.func_info[i].arg_tys.iter().cloned()),
+            );
+            args.reverse();
+            for (a_reg, a_ty) in args.drain(..) {
+                pop_var(instrs, a_reg, a_ty)?;
+            }
+
+            for (j, n) in frame.cfg.raw_nodes().iter().enumerate() {
+                bb_map.push(instrs.len());
+                use HighLevel::*;
+                for stmt in n.weight.iter() {
+                    match stmt {
+                        Either::Left(ll) => instrs.push(ll.clone()),
+                        Either::Right(Call {
+                            func_id,
+                            dst_reg,
+                            dst_ty,
+                            args,
+                        }) => {
+                            // args have already been normalized, and return type already matches.
+                            // All we need to do is push local variables (to avoid clobbers) and
+                            // push args onto the stack.
+                            // NB locals does not contain all of the local registers, though the
+                            // ones it does not cover are "transient" in that we have no way to get
+                            // a handle on them outside of the immediate context in which they are
+                            // constructed (e.g. through reg_of_ty). I believe that means we can
+                            // rule them out as being needed across callsites.
+                            //
+                            // Today, function calls are a bit slow because of all these pushes (we
+                            // have a lot of local variables because we do not reuse registers). We
+                            // may want to optimize this by looking only over variables referenced
+                            // in reachable BBs from the current one.
+                            locals.clear();
+                            locals.extend(frame.locals.values().cloned());
+                            for (reg, ty) in locals.iter().cloned() {
+                                push_var(instrs, reg, ty)?;
+                            }
+                            for (reg, ty) in args.iter().cloned() {
+                                push_var(instrs, reg, ty)?;
+                            }
+                            let callee = *func_id as usize;
+                            instrs.push(LL::Call(callee));
+
+                            // Restore local variables
+                            locals.reverse();
+                            for (reg, ty) in locals.iter().cloned() {
+                                pop_var(instrs, reg, ty)?;
+                            }
+                            let ret_reg = ret_regs[callee];
+                            debug_assert_eq!(self.func_info[callee].ret_ty, *dst_ty);
+                            if let Some(inst) = mov(*dst_reg, ret_reg, *dst_ty)? {
+                                instrs.push(inst);
+                            }
+                        }
+                        Either::Right(Ret(reg, ty)) => {
+                            debug_assert_eq!(self.func_info[i].ret_ty, *ty);
+                            if let Some(inst) = mov(ret_regs[i], *reg, *ty)? {
+                                instrs.push(inst);
+                            }
+                            instrs.push(LL::Ret);
+                        }
+                        // handles by the predecessor.
+                        Either::Right(Phi(_, _, _)) => {}
+                    }
+                }
+
+                let ix = NodeIx::new(j);
+                // Now handle phi nodes
+                for neigh in frame.cfg.neighbors(ix) {
+                    for stmt in frame.cfg.node_weight(neigh).unwrap() {
+                        if let Either::Right(Phi(reg, ty, preds)) = stmt {
+                            for (pred, src_reg) in preds.iter() {
+                                if pred == &ix {
+                                    if let Some(inst) = mov(*reg, *src_reg, *ty)? {
+                                        instrs.push(inst);
+                                    }
+                                    break;
+                                }
+                            }
+                        } else {
+                            // Phis are all at the top;
+                            break;
+                        }
+                    }
+                }
+                // And then jumps
+                let mut walker = frame.cfg.neighbors(ix).detach();
+                let mut edges = SmallVec::new();
+                while let Some(eix) = walker.next_edge(&frame.cfg) {
+                    edges.push(eix)
+                }
+                edges.reverse();
+                for eix in edges.iter().cloned() {
+                    let dst = frame.cfg.edge_endpoints(eix).unwrap().1.index();
+                    if let Some(reg) = frame.cfg.edge_weight(eix).unwrap().clone() {
+                        jmps.push(instrs.len());
+                        instrs.push(LL::JmpIf(reg.into(), dst.into()));
+                    } else if dst != j + 1 {
+                        jmps.push(instrs.len());
+                        instrs.push(LL::Jmp(dst.into()));
+                    }
+                }
+            }
+            // Now rewrite jumps
+            for j in jmps.iter() {
+                match &mut instrs[*j] {
+                    LL::Jmp(bb) | LL::JmpIf(_, bb) => *bb = bb_map[bb.0 as usize].into(),
+                    _ => unreachable!(),
+                }
+            }
+        }
+        Ok(res)
+    }
+
+    fn init_from_ctx(pc: &ProgramContext<'a, &'a str>) -> Result<Typer<'a>> {
+        // Type-check the code, then initialize a Typer, assigning registers to local
         // and global variables.
 
-        let mut gen = ProgramGenerator::default();
+        let mut gen = Typer::default();
         let types::TypeInfo { var_tys, func_tys } = types::get_types(pc)?;
         macro_rules! init_entry {
             ($v:expr, $func_id:expr, $args:expr) => {
                 let res = gen.frames.len() as NumTy;
                 $v.insert(res);
                 let ret_ty = func_tys[&($func_id, $args.clone())];
-                let ret_reg = reg_of_ty!(gen.regs, ret_ty, RegStatus::Ret);
                 let mut f = Frame::default();
                 f.src_function = $func_id;
                 f.cur_ident = res;
                 gen.frames.push(f);
                 gen.func_info.push(FuncInfo {
-                    ret_reg,
                     ret_ty,
-                    args: $args.clone(),
+                    // Want to allocate registers for args here, $args just contains the types.
+                    arg_tys: $args.clone(),
                 });
             };
         }
@@ -209,14 +449,13 @@ impl<'a> ProgramGenerator<'a> {
                 }
                 &mut gen.frames[gen.id_map[&(*func_id, args.clone())] as usize].locals
             };
-            let reg = reg_of_ty!(
-                gen.regs,
+            let reg = gen.regs.stats.new_reg(
                 *ty,
                 if id.global {
                     RegStatus::Global
                 } else {
                     RegStatus::Local
-                }
+                },
             );
             if let Some(old) = map.insert(*id, (reg, *ty)) {
                 return err!(
@@ -231,241 +470,76 @@ impl<'a> ProgramGenerator<'a> {
         gen.main_offset = main_offset as usize;
         for frame in gen.frames.iter_mut() {
             let src_func = frame.src_function as usize;
+            let mut stream = Vec::new();
             View {
                 frame,
                 regs: &mut gen.regs,
                 id_map: &gen.id_map,
                 arity: &gen.arity,
                 func_info: &gen.func_info,
+                stream: &mut stream,
             }
             .process_function(&pc.funcs[src_func])?;
         }
         Ok(gen)
     }
-
-    fn var_refs(
-        &self,
-        // do we want to monomorphize this? I guess not, because we only care about globals.
-    ) -> Vec<
-        /* indexed by byte-code level function */
-        /* globals */ HashSet<(NumTy, Ty)>,
-    > {
-        let mut res = Vec::new();
-        for (i, frame) in self.frames.iter().enumerate() {
-            let mut globals = HashSet::new();
-            let mut f = |reg, ty| {
-                if let RegStatus::Global = self.regs.reg_stats.get_status(reg, ty) {
-                    globals.insert((reg, ty));
-                }
-            };
-            // TODO: make this generic, put it in bytecode or elsewhere
-            // TODO: figure out a story for args (need to map them to locals; maybe make them a
-            // separate register status?)
-            // TODO: create mapping for basic blocks
-            for instr in frame.instrs.iter() {
-                use Instr::*;
-                instr.accum(&mut f);
-            }
-            res.push(globals);
-        }
-        res
-    }
-
-    fn into_interp(
-        mut self,
-        reader: impl std::io::Read + 'static,
-        writer: impl std::io::Write + 'static,
-    ) -> Interp<'a> {
-        let instrs: Vec<_> = self.frames.drain(..).map(|x| x.instrs).collect();
-        Interp::new(
-            instrs,
-            self.main_offset,
-            |ty| self.regs.reg_stats.count(ty) as usize,
-            reader,
-            writer,
-        )
-    }
-}
-
-#[derive(Default)]
-struct Registers {
-    reg_stats: RegStatuses,
-    globals: HashMap<Ident, (u32, Ty)>,
-}
-
-#[derive(Default)]
-struct Frame<'a> {
-    src_function: NumTy,
-    cur_ident: NumTy,
-    locals: HashMap<Ident, (u32, Ty)>,
-    jmps: Vec<usize>,
-    bb_to_instr: Vec<usize>,
-    instrs: Vec<Instr<'a>>,
-
-    // Used for codegen only.
-    calls: HashMap<
-        usize, /* index into instrs*/
-        (
-            NumTy, /*return reg*/
-            Ty,
-            SmallVec<(NumTy /*arg reg*/, Ty)>,
-        ),
-    >,
-    rets: HashMap<usize, Option<(NumTy /*return reg*/, Ty)>>,
-}
-
-struct View<'a, 'b> {
-    frame: &'b mut Frame<'a>,
-    regs: &'b mut Registers,
-    id_map: &'b HashMap<(NumTy, SmallVec<Ty>), NumTy>,
-    arity: &'b HashMap<NumTy, NumTy>,
-    func_info: &'b Vec<FuncInfo>,
 }
 
 impl<'a, 'b> View<'a, 'b> {
-    fn push_var(&mut self, reg: NumTy, ty: Ty) -> Result<()> {
-        use Ty::*;
-        self.push(match ty {
-            Int => Instr::PushInt(reg.into()),
-            Float => Instr::PushFloat(reg.into()),
-            Str => Instr::PushStr(reg.into()),
-            MapIntInt => Instr::PushIntInt(reg.into()),
-            MapIntFloat => Instr::PushIntFloat(reg.into()),
-            MapIntStr => Instr::PushIntStr(reg.into()),
-            MapStrInt => Instr::PushStrInt(reg.into()),
-            MapStrFloat => Instr::PushStrFloat(reg.into()),
-            MapStrStr => Instr::PushStrStr(reg.into()),
-            IterInt | IterStr => return err!("invalid argument type: {:?}", ty),
-        });
-        Ok(())
-    }
-    fn pop_var(&mut self, reg: NumTy, ty: Ty) -> Result<()> {
-        use Ty::*;
-        self.push(match ty {
-            Int => Instr::PopInt(reg.into()),
-            Float => Instr::PopFloat(reg.into()),
-            Str => Instr::PopStr(reg.into()),
-            MapIntInt => Instr::PopIntInt(reg.into()),
-            MapIntFloat => Instr::PopIntFloat(reg.into()),
-            MapIntStr => Instr::PopIntStr(reg.into()),
-            MapStrInt => Instr::PopStrInt(reg.into()),
-            MapStrFloat => Instr::PopStrFloat(reg.into()),
-            MapStrStr => Instr::PopStrStr(reg.into()),
-            IterInt | IterStr => return err!("invalid argument type: {:?}", ty),
-        });
-        Ok(())
-    }
     fn process_function(&mut self, func: &Function<'a, &'a str>) -> Result<()> {
-        // No logic in this method relies on this directly, but it is an
-        // invariant of the broader module.
-        assert_eq!(self.frame.src_function, func.ident);
-
-        // Pop any arguments off the stack in reverse order.
-        for arg in func.args.iter().rev() {
-            let (reg, ty) = self.reg_of_ident(&arg.id);
-            self.pop_var(reg, ty)?;
+        // Record registers for arguments.
+        for arg in func.args.iter() {
+            let (reg, _) = self.reg_of_ident(&arg.id);
+            self.frame.arg_regs.push(reg);
         }
-        self.frame.bb_to_instr = vec![0; func.cfg.node_count()];
-
+        // Allocate basic blocks in CFG.
+        for _ in 0..func.cfg.node_count() {
+            self.frame.cfg.add_node(Default::default());
+        }
+        // Fill them in.
         for (i, n) in func.cfg.raw_nodes().iter().enumerate() {
-            self.frame.bb_to_instr[i] = self.frame.instrs.len();
             for stmt in n.weight.q.iter() {
                 self.stmt(stmt)?;
             }
-
             let ix = NodeIx::new(i);
-            let mut branches: cfg::SmallVec<petgraph::graph::EdgeIndex> = Default::default();
+            let mut branches: SmallVec<petgraph::graph::EdgeIndex> = Default::default();
             let mut walker = func.cfg.neighbors(ix).detach();
             while let Some(e) = walker.next_edge(&func.cfg) {
-                branches.push(e)
+                branches.push(e);
             }
-
-            // Petgraph gives us edges back in reverse order.
+            // We get branches back in reverse order.
             branches.reverse();
-
-            // Replace Phi functions in successors with assignments at this point in the stream.
-            //
-            // NB Why is it sufficient to do all assignments here? Shouldn't we limit code gen to  the
-            // branch we are actually going to take? No, because one must first go through another
-            // block that assigns to the node again before actually reading the variable.
-            for n in func.cfg.neighbors(NodeIx::new(i)) {
-                let weight = func.cfg.node_weight(n).unwrap();
-                for stmt in weight.q.iter() {
-                    if let PrimStmt::AsgnVar(id, PrimExpr::Phi(preds)) = stmt {
-                        let mut found = false;
-                        for (pred, src) in preds.iter() {
-                            if pred.index() == i {
-                                found = true;
-                                // now do the assignment
-                                let (dst_reg, dst_ty) = self.reg_of_ident(id);
-                                let (src_reg, src_ty) = self.reg_of_ident(src);
-                                self.convert(dst_reg, dst_ty, src_reg, src_ty)?;
-                            }
-                        }
-                        if !found {
-                            return err!("malformed phi node: preds={:?} cur={:?}", &preds[..], i);
-                        }
-                    } else {
-                        break;
-                    }
-                }
-            }
-
-            // Insert code for the branches at the end of this basic block. At first, labels just
-            // indicate the basic block in the CFG. They are re-mapped at the end of execution.
-            // NB Don't insert a halt if this block ends with a return.
-            let mut is_end = i != func.exit.index();
-            for b in branches.iter().cloned() {
-                let next = func.cfg.edge_endpoints(b).unwrap().1;
-                match &func.cfg.edge_weight(b).unwrap().0 {
-                    Some(v) => {
-                        let (mut reg, ty, status) = self.get_reg_status(v)?;
+            for eix in branches.iter().cloned() {
+                let transition = func.cfg.edge_weight(eix).unwrap();
+                let edge = match &transition.0 {
+                    Some(val) => {
+                        let (mut reg, ty) = self.get_reg(val)?;
                         match ty {
                             Ty::Int => {}
                             Ty::Float => {
-                                let dst = reg_of_ty!(self.regs, Ty::Int, status);
+                                let dst = self.regs.stats.reg_of_ty(Ty::Int);
                                 self.convert(dst, Ty::Int, reg, Ty::Float)?;
                                 reg = dst;
                             }
                             Ty::Str => {
-                                let dst = reg_of_ty!(self.regs, Ty::Int, status);
-                                self.push(Instr::LenStr(dst.into(), reg.into()));
+                                let dst = self.regs.stats.reg_of_ty(Ty::Int);
+                                self.pushl(LL::LenStr(dst.into(), reg.into()));
                                 reg = dst;
                             }
-                            _ => return err!("invalid type for branch: {:?}: {:?}", v, ty),
+                            _ => return err!("invalid type for branch: {:?} :: {:?}", val, ty),
                         }
-
-                        let end_pos = self.frame.instrs.len();
-                        self.frame.jmps.push(end_pos);
-                        self.push(Instr::JmpIf(reg.into(), next.index().into()));
+                        Some(reg)
                     }
-                    None => {
-                        is_end = false;
-                        // There's no point issuing an unconditional jump to the next basic block
-                        // because it comes next in the instruction stream.
-                        if next.index() != i + 1 {
-                            self.frame.jmps.push(self.frame.instrs.len());
-                            self.push(Instr::Jmp(next.index().into()))
-                        }
-                    }
-                }
+                    None => None,
+                };
+                let (src, dst) = func.cfg.edge_endpoints(eix).unwrap();
+                self.frame.cfg.add_edge(src, dst, edge);
             }
-            if is_end {
-                // TODO Halt is no longer emitted; we may want to remove it now that there's
-                // Return.
-                self.push(Instr::Halt);
-            }
-        }
-
-        // Rewrite jumps to point to proper offsets.
-        for jmp in self.frame.jmps.iter().cloned() {
-            match &mut self.frame.instrs[jmp] {
-                Instr::Jmp(bytecode::Label(l)) => *l = self.frame.bb_to_instr[*l as usize] as u32,
-                Instr::JmpIf(_, bytecode::Label(l)) => {
-                    *l = self.frame.bb_to_instr[*l as usize] as u32
-                }
-                _ => unreachable!(),
-            }
+            // In the interim, someone may have added some instructions to our basic block when
+            // processing a Phi function. Merge in any of those changes.
+            let cur_bb = self.frame.cfg.node_weight_mut(ix).unwrap();
+            self.stream.extend(cur_bb.drain(..));
+            mem::swap(cur_bb, self.stream);
         }
         Ok(())
     }
@@ -476,6 +550,7 @@ impl<'a, 'b> View<'a, 'b> {
         let (reg, ty, _status) = self.reg_of_ident_status(id);
         (reg, ty)
     }
+
     fn reg_of_ident_status(&mut self, id: &Ident) -> (u32, Ty, RegStatus) {
         if is_unused(*id) {
             // We should not actually store into the "unused" identifier.
@@ -488,36 +563,15 @@ impl<'a, 'b> View<'a, 'b> {
         } else {
             (self.frame.locals[id], RegStatus::Local)
         };
-        dbg_reg_valid!(self, res_reg, res_ty);
         (res_reg, res_ty, status)
     }
 
-    // Move src into dst at type Ty.
-    fn mov(&mut self, dst_reg: u32, src_reg: u32, ty: Ty) -> Result<()> {
-        use Ty::*;
-        if dst_reg == UNUSED || src_reg == UNUSED {
-            return Ok(());
-        }
+    fn pushl(&mut self, i: LL<'a>) {
+        self.stream.push(Either::Left(i))
+    }
 
-        dbg_reg_valid!(self, dst_reg, ty);
-        dbg_reg_valid!(self, src_reg, ty);
-
-        let res = match ty {
-            Int => Instr::MovInt(dst_reg.into(), src_reg.into()),
-            Float => Instr::MovFloat(dst_reg.into(), src_reg.into()),
-            Str => Instr::MovStr(dst_reg.into(), src_reg.into()),
-
-            MapIntInt => Instr::MovMapIntInt(dst_reg.into(), src_reg.into()),
-            MapIntFloat => Instr::MovMapIntFloat(dst_reg.into(), src_reg.into()),
-            MapIntStr => Instr::MovMapIntStr(dst_reg.into(), src_reg.into()),
-
-            MapStrInt => Instr::MovMapStrInt(dst_reg.into(), src_reg.into()),
-            MapStrFloat => Instr::MovMapStrFloat(dst_reg.into(), src_reg.into()),
-            MapStrStr => Instr::MovMapStrStr(dst_reg.into(), src_reg.into()),
-
-            IterInt | IterStr => return err!("attempt to move values of type {:?}", ty),
-        };
-        Ok(self.push(res))
+    fn pushr(&mut self, i: HighLevel) {
+        self.stream.push(Either::Right(i))
     }
 
     // Get the register associated with a value. For identifiers this has the same semantics as
@@ -531,26 +585,32 @@ impl<'a, 'b> View<'a, 'b> {
         use RegStatus::*;
         match v {
             PrimVal::ILit(i) => {
-                let nreg = reg_of_ty!(self.regs, Ty::Int, Local);
-                self.push(Instr::StoreConstInt(nreg.into(), *i));
+                let nreg = self.regs.stats.new_reg(Ty::Int, Local);
+                self.pushl(LL::StoreConstInt(nreg.into(), *i));
                 Ok((nreg, Ty::Int, Local))
             }
             PrimVal::FLit(f) => {
-                let nreg = reg_of_ty!(self.regs, Ty::Float, Local);
-                self.push(Instr::StoreConstFloat(nreg.into(), *f));
+                let nreg = self.regs.stats.new_reg(Ty::Float, Local);
+                self.pushl(LL::StoreConstFloat(nreg.into(), *f));
                 Ok((nreg, Ty::Float, Local))
             }
             PrimVal::StrLit(s) => {
-                let nreg = reg_of_ty!(self.regs, Ty::Str, Local);
-                self.push(Instr::StoreConstStr(nreg.into(), (*s).into()));
+                let nreg = self.regs.stats.new_reg(Ty::Str, Local);
+                self.pushl(LL::StoreConstStr(nreg.into(), (*s).into()));
                 Ok((nreg, Ty::Str, Local))
             }
             PrimVal::Var(v) => Ok(self.reg_of_ident_status(v)),
         }
     }
 
-    // Convert src into dst. If the types match this is just a move. If both the types and the
-    // registers match this is a noop.
+    // Move src into dst at type Ty.
+    fn mov(&mut self, dst_reg: u32, src_reg: u32, ty: Ty) -> Result<()> {
+        if let Some(inst) = mov(dst_reg, src_reg, ty)? {
+            self.pushl(inst);
+        }
+        Ok(())
+    }
+
     fn convert(&mut self, dst_reg: u32, dst_ty: Ty, src_reg: u32, src_ty: Ty) -> Result<()> {
         use Ty::*;
         if dst_reg == UNUSED || src_reg == UNUSED {
@@ -559,24 +619,22 @@ impl<'a, 'b> View<'a, 'b> {
         if dst_reg == src_reg && dst_ty == src_ty {
             return Ok(());
         }
-        dbg_reg_valid!(self, dst_reg, dst_ty);
-        dbg_reg_valid!(self, src_reg, src_ty);
 
         let res = match (dst_ty, src_ty) {
-            (Float, Int) => Instr::IntToFloat(dst_reg.into(), src_reg.into()),
-            (Str, Int) => Instr::IntToStr(dst_reg.into(), src_reg.into()),
+            (Float, Int) => LL::IntToFloat(dst_reg.into(), src_reg.into()),
+            (Str, Int) => LL::IntToStr(dst_reg.into(), src_reg.into()),
 
-            (Int, Float) => Instr::FloatToInt(dst_reg.into(), src_reg.into()),
-            (Str, Float) => Instr::FloatToStr(dst_reg.into(), src_reg.into()),
+            (Int, Float) => LL::FloatToInt(dst_reg.into(), src_reg.into()),
+            (Str, Float) => LL::FloatToStr(dst_reg.into(), src_reg.into()),
 
-            (Int, Str) => Instr::StrToInt(dst_reg.into(), src_reg.into()),
-            (Float, Str) => Instr::StrToFloat(dst_reg.into(), src_reg.into()),
+            (Int, Str) => LL::StrToInt(dst_reg.into(), src_reg.into()),
+            (Float, Str) => LL::StrToFloat(dst_reg.into(), src_reg.into()),
 
-            (MapIntFloat, MapIntFloat) => Instr::MovMapIntFloat(dst_reg.into(), src_reg.into()),
-            (MapIntStr, MapIntStr) => Instr::MovMapIntStr(dst_reg.into(), src_reg.into()),
+            (MapIntFloat, MapIntFloat) => LL::MovMapIntFloat(dst_reg.into(), src_reg.into()),
+            (MapIntStr, MapIntStr) => LL::MovMapIntStr(dst_reg.into(), src_reg.into()),
 
-            (MapStrFloat, MapStrFloat) => Instr::MovMapStrFloat(dst_reg.into(), src_reg.into()),
-            (MapStrStr, MapStrStr) => Instr::MovMapStrStr(dst_reg.into(), src_reg.into()),
+            (MapStrFloat, MapStrFloat) => LL::MovMapStrFloat(dst_reg.into(), src_reg.into()),
+            (MapStrStr, MapStrStr) => LL::MovMapStrStr(dst_reg.into(), src_reg.into()),
             (dst, src) => {
                 if dst == src {
                     return self.mov(dst_reg, src_reg, dst);
@@ -589,41 +647,40 @@ impl<'a, 'b> View<'a, 'b> {
             }
         };
 
-        Ok(self.push(res))
+        Ok(self.pushl(res))
     }
 
     // Store values into a register at a given type, converting if necessary.
     fn store(&mut self, dst_reg: u32, dst_ty: Ty, src: &PrimVal<'a>) -> Result<()> {
-        dbg_reg_valid!(self, dst_reg, dst_ty);
         match src {
-            PrimVal::Var(id2) => {
-                let (src_reg, src_ty) = self.reg_of_ident(id2);
+            PrimVal::Var(id) => {
+                let (src_reg, src_ty) = self.reg_of_ident(id);
                 self.convert(dst_reg, dst_ty, src_reg, src_ty)?;
             }
             PrimVal::ILit(i) => {
                 if dst_ty == Ty::Int {
-                    self.push(Instr::StoreConstInt(dst_reg.into(), *i));
+                    self.pushl(LL::StoreConstInt(dst_reg.into(), *i));
                 } else {
-                    let ir = reg_of_ty!(self.regs, Ty::Int);
-                    self.push(Instr::StoreConstInt(ir.into(), *i));
+                    let ir = self.regs.stats.reg_of_ty(Ty::Int);
+                    self.pushl(LL::StoreConstInt(ir.into(), *i));
                     self.convert(dst_reg, dst_ty, ir, Ty::Int)?;
                 }
             }
             PrimVal::FLit(f) => {
                 if dst_ty == Ty::Float {
-                    self.push(Instr::StoreConstFloat(dst_reg.into(), *f));
+                    self.pushl(LL::StoreConstFloat(dst_reg.into(), *f));
                 } else {
-                    let ir = reg_of_ty!(self.regs, Ty::Float);
-                    self.push(Instr::StoreConstFloat(ir.into(), *f));
+                    let ir = self.regs.stats.reg_of_ty(Ty::Float);
+                    self.pushl(LL::StoreConstFloat(ir.into(), *f));
                     self.convert(dst_reg, dst_ty, ir, Ty::Float)?;
                 }
             }
             PrimVal::StrLit(s) => {
                 if dst_ty == Ty::Str {
-                    self.push(Instr::StoreConstStr(dst_reg.into(), (*s).into()));
+                    self.pushl(LL::StoreConstStr(dst_reg.into(), (*s).into()));
                 } else {
-                    let ir = reg_of_ty!(self.regs, Ty::Str);
-                    self.push(Instr::StoreConstStr(ir.into(), (*s).into()));
+                    let ir = self.regs.stats.reg_of_ty(Ty::Str);
+                    self.pushl(LL::StoreConstStr(ir.into(), (*s).into()));
                     self.convert(dst_reg, dst_ty, ir, Ty::Str)?;
                 }
             }
@@ -643,13 +700,11 @@ impl<'a, 'b> View<'a, 'b> {
         if dst_reg == UNUSED {
             return Ok(());
         }
-        dbg_reg_valid!(self, dst_reg, dst_ty);
-        dbg_reg_valid!(self, arr_reg, arr_ty);
         // Convert `key` if necessary.
         let target_ty = arr_ty.key()?;
         let (mut key_reg, key_ty) = self.get_reg(key)?;
         if target_ty != key_ty {
-            let inter = reg_of_ty!(self.regs, target_ty);
+            let inter = self.regs.stats.reg_of_ty(target_ty);
             self.convert(inter, target_ty, key_reg, key_ty)?;
             key_reg = inter;
         }
@@ -659,18 +714,18 @@ impl<'a, 'b> View<'a, 'b> {
         let load_reg = if dst_ty == arr_val_ty {
             dst_reg
         } else {
-            reg_of_ty!(self.regs, arr_val_ty)
+            self.regs.stats.reg_of_ty(arr_val_ty)
         };
 
         // Emit the corresponding instruction.
         use Ty::*;
-        self.push(match arr_ty {
-            MapIntInt => Instr::LookupIntInt(load_reg.into(), arr_reg.into(), key_reg.into()),
-            MapIntFloat => Instr::LookupIntFloat(load_reg.into(), arr_reg.into(), key_reg.into()),
-            MapIntStr => Instr::LookupIntStr(load_reg.into(), arr_reg.into(), key_reg.into()),
-            MapStrInt => Instr::LookupStrInt(load_reg.into(), arr_reg.into(), key_reg.into()),
-            MapStrFloat => Instr::LookupStrFloat(load_reg.into(), arr_reg.into(), key_reg.into()),
-            MapStrStr => Instr::LookupStrStr(load_reg.into(), arr_reg.into(), key_reg.into()),
+        self.pushl(match arr_ty {
+            MapIntInt => LL::LookupIntInt(load_reg.into(), arr_reg.into(), key_reg.into()),
+            MapIntFloat => LL::LookupIntFloat(load_reg.into(), arr_reg.into(), key_reg.into()),
+            MapIntStr => LL::LookupIntStr(load_reg.into(), arr_reg.into(), key_reg.into()),
+            MapStrInt => LL::LookupStrInt(load_reg.into(), arr_reg.into(), key_reg.into()),
+            MapStrFloat => LL::LookupStrFloat(load_reg.into(), arr_reg.into(), key_reg.into()),
+            MapStrStr => LL::LookupStrStr(load_reg.into(), arr_reg.into(), key_reg.into()),
             Int | Float | Str | IterInt | IterStr => {
                 return err!("[load_map] expected map type, found {:?}", arr_ty)
             }
@@ -678,6 +733,7 @@ impl<'a, 'b> View<'a, 'b> {
         // Convert the result: note that if we had load_reg == dst_reg, then this is a noop.
         self.convert(dst_reg, dst_ty, load_reg, arr_val_ty)
     }
+
     fn builtin(
         &mut self,
         dst_reg: u32,
@@ -696,7 +752,6 @@ impl<'a, 'b> View<'a, 'b> {
             args_regs.push(reg);
             args_tys.push(ty);
         }
-        dbg_reg_valid!(self, dst_reg, dst_ty);
 
         // Now, perform any necessary conversions if input types do not match the argument types.
         let mut conv_regs: cfg::SmallVec<_> = smallvec![UNUSED; args.len()];
@@ -711,7 +766,7 @@ impl<'a, 'b> View<'a, 'b> {
             if aty == cty {
                 *creg = areg;
             } else {
-                let reg = reg_of_ty!(self.regs, cty);
+                let reg = self.regs.stats.reg_of_ty(cty);
                 self.convert(reg, cty, areg, aty)?;
                 *creg = reg;
             }
@@ -720,10 +775,8 @@ impl<'a, 'b> View<'a, 'b> {
         let mut res_reg = if dst_ty == res_ty {
             dst_reg
         } else {
-            reg_of_ty!(self.regs, res_ty)
+            self.regs.stats.reg_of_ty(res_ty)
         };
-
-        dbg_reg_valid!(self, res_reg, res_ty);
 
         // Helper macro for generating code for binary operators
         macro_rules! gen_op {
@@ -731,7 +784,7 @@ impl<'a, 'b> View<'a, 'b> {
                 match conv_tys[0] {
                     // TODO implement better handling of "NULL" and get rid of hacks like this.
                     $( Ty::$ty  => if res_reg != UNUSED {
-                        self.push(Instr::$inst(res_reg.into(),
+                        self.pushl(LL::$inst(res_reg.into(),
                                         conv_regs[0].into(),
                                         conv_regs[1].into()));
                     }, )*
@@ -741,17 +794,17 @@ impl<'a, 'b> View<'a, 'b> {
         };
 
         match bf {
-            Unop(Column) => self.push(Instr::GetColumn(res_reg.into(), conv_regs[0].into())),
-            Unop(Not) => self.push(if conv_tys[0] == Ty::Str {
-                Instr::NotStr(res_reg.into(), conv_regs[0].into())
+            Unop(Column) => self.pushl(LL::GetColumn(res_reg.into(), conv_regs[0].into())),
+            Unop(Not) => self.pushl(if conv_tys[0] == Ty::Str {
+                LL::NotStr(res_reg.into(), conv_regs[0].into())
             } else {
                 debug_assert_eq!(conv_tys[0], Ty::Int);
-                Instr::Not(res_reg.into(), conv_regs[0].into())
+                LL::Not(res_reg.into(), conv_regs[0].into())
             }),
-            Unop(Neg) => self.push(if conv_tys[0] == Ty::Float {
-                Instr::NegFloat(res_reg.into(), conv_regs[0].into())
+            Unop(Neg) => self.pushl(if conv_tys[0] == Ty::Float {
+                LL::NegFloat(res_reg.into(), conv_regs[0].into())
             } else {
-                Instr::NegInt(res_reg.into(), conv_regs[0].into())
+                LL::NegInt(res_reg.into(), conv_regs[0].into())
             }),
             Unop(Pos) => self.mov(res_reg, conv_regs[0], conv_tys[0])?,
             Binop(Plus) => gen_op!(Plus, [Float, AddFloat], [Int, AddInt]),
@@ -775,24 +828,24 @@ impl<'a, 'b> View<'a, 'b> {
                 [MapStrStr, ContainsStrStr],
                 [MapStrFloat, ContainsStrFloat]
             ),
-            ReadErr => self.push(Instr::ReadErr(res_reg.into(), conv_regs[0].into())),
-            Nextline => self.push(Instr::NextLine(res_reg.into(), conv_regs[0].into())),
-            ReadErrStdin => self.push(Instr::ReadErrStdin(res_reg.into())),
-            NextlineStdin => self.push(Instr::NextLineStdin(res_reg.into())),
-            Setcol => self.push(Instr::SetColumn(conv_regs[0].into(), conv_regs[1].into())),
+            ReadErr => self.pushl(LL::ReadErr(res_reg.into(), conv_regs[0].into())),
+            Nextline => self.pushl(LL::NextLine(res_reg.into(), conv_regs[0].into())),
+            ReadErrStdin => self.pushl(LL::ReadErrStdin(res_reg.into())),
+            NextlineStdin => self.pushl(LL::NextLineStdin(res_reg.into())),
+            Setcol => self.pushl(LL::SetColumn(conv_regs[0].into(), conv_regs[1].into())),
             Split => {
                 if res_reg == UNUSED {
-                    res_reg = reg_of_ty!(self.regs, res_ty);
+                    res_reg = self.regs.stats.reg_of_ty(res_ty);
                 }
-                self.push(if conv_tys[1] == Ty::MapIntStr {
-                    Instr::SplitInt(
+                self.pushl(if conv_tys[1] == Ty::MapIntStr {
+                    LL::SplitInt(
                         res_reg.into(),
                         conv_regs[0].into(),
                         conv_regs[1].into(),
                         conv_regs[2].into(),
                     )
                 } else if conv_tys[1] == Ty::MapStrStr {
-                    Instr::SplitStr(
+                    LL::SplitStr(
                         res_reg.into(),
                         conv_regs[0].into(),
                         conv_regs[1].into(),
@@ -802,59 +855,49 @@ impl<'a, 'b> View<'a, 'b> {
                     return err!("invalid input types to split: {:?}", &conv_tys[..]);
                 })
             }
-            Length => self.push(match conv_tys[0] {
-                Ty::MapIntInt => Instr::LenIntInt(res_reg.into(), conv_regs[0].into()),
-                Ty::MapIntStr => Instr::LenIntStr(res_reg.into(), conv_regs[0].into()),
-                Ty::MapIntFloat => Instr::LenIntFloat(res_reg.into(), conv_regs[0].into()),
-                Ty::MapStrInt => Instr::LenStrInt(res_reg.into(), conv_regs[0].into()),
-                Ty::MapStrStr => Instr::LenStrStr(res_reg.into(), conv_regs[0].into()),
-                Ty::MapStrFloat => Instr::LenStrFloat(res_reg.into(), conv_regs[0].into()),
-                Ty::Str => Instr::LenStr(res_reg.into(), conv_regs[0].into()),
+            Length => self.pushl(match conv_tys[0] {
+                Ty::MapIntInt => LL::LenIntInt(res_reg.into(), conv_regs[0].into()),
+                Ty::MapIntStr => LL::LenIntStr(res_reg.into(), conv_regs[0].into()),
+                Ty::MapIntFloat => LL::LenIntFloat(res_reg.into(), conv_regs[0].into()),
+                Ty::MapStrInt => LL::LenStrInt(res_reg.into(), conv_regs[0].into()),
+                Ty::MapStrStr => LL::LenStrStr(res_reg.into(), conv_regs[0].into()),
+                Ty::MapStrFloat => LL::LenStrFloat(res_reg.into(), conv_regs[0].into()),
+                Ty::Str => LL::LenStr(res_reg.into(), conv_regs[0].into()),
                 _ => return err!("invalid input type for length: {:?}", &conv_tys[..]),
             }),
             Print => {
                 // XXX this imports a specific assumption on how the PrimStmt is generated, we may
                 // want to make the bool parameter to Print dynamic.
                 if let cfg::PrimVal::ILit(i) = &args[2] {
-                    self.push(Instr::Print(
-                        conv_regs[0].into(),
-                        conv_regs[1].into(),
-                        *i != 0,
-                    ));
+                    self.pushl(LL::Print(conv_regs[0].into(), conv_regs[1].into(), *i != 0));
                     return Ok(());
                 } else {
                     return err!("must pass constant append parameter to print");
                 }
             }
             PrintStdout => {
-                self.push(Instr::PrintStdout(conv_regs[0].into()));
+                self.pushl(LL::PrintStdout(conv_regs[0].into()));
                 return Ok(());
             }
             Delete => match &conv_tys[0] {
-                Ty::MapIntInt => self.push(Instr::DeleteIntInt(
-                    conv_regs[0].into(),
-                    conv_regs[1].into(),
-                )),
-                Ty::MapIntStr => self.push(Instr::DeleteIntStr(
-                    conv_regs[0].into(),
-                    conv_regs[1].into(),
-                )),
-                Ty::MapIntFloat => self.push(Instr::DeleteIntFloat(
-                    conv_regs[0].into(),
-                    conv_regs[1].into(),
-                )),
-                Ty::MapStrInt => self.push(Instr::DeleteStrInt(
-                    conv_regs[0].into(),
-                    conv_regs[1].into(),
-                )),
-                Ty::MapStrStr => self.push(Instr::DeleteStrStr(
-                    conv_regs[0].into(),
-                    conv_regs[1].into(),
-                )),
-                Ty::MapStrFloat => self.push(Instr::DeleteStrFloat(
-                    conv_regs[0].into(),
-                    conv_regs[1].into(),
-                )),
+                Ty::MapIntInt => {
+                    self.pushl(LL::DeleteIntInt(conv_regs[0].into(), conv_regs[1].into()))
+                }
+                Ty::MapIntStr => {
+                    self.pushl(LL::DeleteIntStr(conv_regs[0].into(), conv_regs[1].into()))
+                }
+                Ty::MapIntFloat => {
+                    self.pushl(LL::DeleteIntFloat(conv_regs[0].into(), conv_regs[1].into()))
+                }
+                Ty::MapStrInt => {
+                    self.pushl(LL::DeleteStrInt(conv_regs[0].into(), conv_regs[1].into()))
+                }
+                Ty::MapStrStr => {
+                    self.pushl(LL::DeleteStrStr(conv_regs[0].into(), conv_regs[1].into()))
+                }
+                Ty::MapStrFloat => {
+                    self.pushl(LL::DeleteStrFloat(conv_regs[0].into(), conv_regs[1].into()))
+                }
                 _ => return err!("incorrect parameter types for Delete: {:?}", &conv_tys[..]),
             },
         };
@@ -864,79 +907,77 @@ impl<'a, 'b> View<'a, 'b> {
     fn expr(&mut self, dst_reg: u32, dst_ty: Ty, exp: &cfg::PrimExpr<'a>) -> Result<()> {
         match exp {
             PrimExpr::Val(v) => self.store(dst_reg, dst_ty, v)?,
-            // Phi functions are handled elsewhere
-            PrimExpr::Phi(_) => {}
+            PrimExpr::Phi(preds) => {
+                let mut pred_regs = SmallVec::with_capacity(preds.len());
+                for (prev, id) in preds.iter().cloned() {
+                    let (id_reg, id_ty) = self.reg_of_ident(&id);
+                    if id_ty == dst_ty {
+                        pred_regs.push((prev, id_reg));
+                        continue;
+                    }
+                    // `id` doesn't have the right type. To handle this, we need to reach into the
+                    // cfg, grab the stream, and append conversion code to it.
+                    //
+                    // We do two loads into cfg for the sake of our friend the borrow checker.
+                    // NB this is mostly required because Phis have to be at the top of
+                    // their basic block. We could probably do the conversions here otherwise.
+                    {
+                        let other_bb = self.frame.cfg.node_weight_mut(prev).unwrap();
+                        mem::swap(self.stream, other_bb);
+                    }
+                    let conv_reg = self.regs.stats.reg_of_ty(dst_ty);
+                    self.convert(conv_reg, dst_ty, id_reg, id_ty)?;
+                    {
+                        let other_bb = self.frame.cfg.node_weight_mut(prev).unwrap();
+                        mem::swap(self.stream, other_bb);
+                    }
+                    pred_regs.push((prev, conv_reg));
+                }
+                self.pushr(HighLevel::Phi(dst_reg, dst_ty, pred_regs));
+            }
             PrimExpr::CallBuiltin(bf, vs) => self.builtin(dst_reg, dst_ty, bf, vs)?,
-            // TODO return register into dst_reg
             PrimExpr::CallUDF(func_id, vs) => {
-                let mut arg_regs: SmallVec<_> = SmallVec::with_capacity(vs.len());
-                let mut arg_tys: SmallVec<_> = SmallVec::with_capacity(vs.len());
+                let mut args = SmallVec::with_capacity(vs.len());
                 for v in vs.iter() {
-                    let (reg, ty) = self.get_reg(v)?;
-                    arg_regs.push(reg);
-                    arg_tys.push(ty);
+                    args.push(self.get_reg(v)?);
                 }
                 // Normalize the call
                 let true_arity = self.arity[func_id] as usize;
-                if arg_regs.len() < true_arity {
+                if args.len() < true_arity {
                     // Not enough arguments; fill in the rest with nulls.
                     // TODO reuse registers here. This is wasteful for functions with a lot of
                     // arguments.
                     let null_ty = types::null_ty();
-                    let null_reg = reg_of_ty!(self.regs, null_ty);
-                    for _ in 0..(true_arity - arg_regs.len()) {
-                        arg_regs.push(null_reg);
-                        arg_tys.push(null_ty);
+                    let null_reg = self.regs.stats.reg_of_ty(null_ty);
+                    for _ in 0..(true_arity - args.len()) {
+                        args.push((null_reg, null_ty));
                     }
                 }
-                if arg_regs.len() > true_arity {
+                if args.len() > true_arity {
                     // Too many arguments; don't pass the extra.
-                    arg_regs.truncate(true_arity);
-                    arg_tys.truncate(true_arity);
+                    args.truncate(true_arity);
                 }
-                let monomorphized = self.id_map[&(*func_id, arg_tys.clone())];
+                let monomorphized =
+                    self.id_map[&(*func_id, args.iter().map(|(_, y)| y).cloned().collect())];
                 let info = &self.func_info[monomorphized as usize];
-                // Save all local variables onto the stack.
-                //
-                // TODO Optimize this to only save the clobbered registers?
-                let mut locals: SmallVec<_> = self.frame.locals.values().cloned().collect();
-                for (reg, ty) in locals.iter().cloned() {
-                    self.push_var(reg, ty)?;
-                }
-                let mut arg_info = SmallVec::with_capacity(arg_regs.len());
-                // Push onto the stack in the given order; we pop them off in reverse in
-                // `process_function`.
-                for (reg, ty) in arg_regs.iter().cloned().zip(arg_tys.iter().cloned()) {
-                    self.push_var(reg, ty)?;
-                    arg_info.push((reg, ty));
-                }
-                let call_index = self.push_with_len(Instr::Call(monomorphized as usize));
 
-                // We've returned. Restore any local variables that we saved on the stack and that
-                // could have gotten clobbered in the meantime.
-                locals.reverse();
-                for (reg, ty) in locals.iter().cloned() {
-                    self.pop_var(reg, ty)?;
-                }
-
-                // Move the function's return value into the destination.
-                // For insertion into into `calls` map.
-                let mut call_reg = dst_reg;
-                let ret_ty = info.ret_ty;
-                let (src_reg, src_ty) = if ret_ty != dst_ty {
-                    // For code generation, we need to have a register in our frame that is the
-                    // same type as the return register. Allocate one if necessary.
-                    let inter_reg = reg_of_ty!(self.regs, ret_ty);
-                    self.mov(inter_reg, info.ret_reg, ret_ty)?;
-                    call_reg = inter_reg;
-                    (inter_reg, ret_ty)
+                if dst_ty != info.ret_ty {
+                    let inter_reg = self.regs.stats.reg_of_ty(info.ret_ty);
+                    self.pushr(HighLevel::Call {
+                        dst_reg: inter_reg,
+                        dst_ty: info.ret_ty,
+                        func_id: monomorphized,
+                        args,
+                    });
+                    self.convert(dst_reg, dst_ty, inter_reg, info.ret_ty)?;
                 } else {
-                    (info.ret_reg, info.ret_ty)
-                };
-                self.convert(dst_reg, dst_ty, src_reg, src_ty)?;
-                self.frame
-                    .calls
-                    .insert(call_index, (call_reg, ret_ty, arg_info));
+                    self.pushr(HighLevel::Call {
+                        dst_reg,
+                        dst_ty,
+                        func_id: monomorphized,
+                        args,
+                    });
+                }
             }
             PrimExpr::Index(arr, k) => {
                 let (arr_reg, arr_ty) = if let PrimVal::Var(arr_id) = arr {
@@ -955,13 +996,13 @@ impl<'a, 'b> View<'a, 'b> {
                         arr_ty
                     );
                 }
-                self.push(match arr_ty {
-                    Ty::MapIntInt => Instr::IterBeginIntInt(dst_reg.into(), arr_reg.into()),
-                    Ty::MapIntFloat => Instr::IterBeginIntFloat(dst_reg.into(), arr_reg.into()),
-                    Ty::MapIntStr => Instr::IterBeginIntStr(dst_reg.into(), arr_reg.into()),
-                    Ty::MapStrInt => Instr::IterBeginStrInt(dst_reg.into(), arr_reg.into()),
-                    Ty::MapStrFloat => Instr::IterBeginStrFloat(dst_reg.into(), arr_reg.into()),
-                    Ty::MapStrStr => Instr::IterBeginStrStr(dst_reg.into(), arr_reg.into()),
+                self.pushl(match arr_ty {
+                    Ty::MapIntInt => LL::IterBeginIntInt(dst_reg.into(), arr_reg.into()),
+                    Ty::MapIntFloat => LL::IterBeginIntFloat(dst_reg.into(), arr_reg.into()),
+                    Ty::MapIntStr => LL::IterBeginIntStr(dst_reg.into(), arr_reg.into()),
+                    Ty::MapStrInt => LL::IterBeginStrInt(dst_reg.into(), arr_reg.into()),
+                    Ty::MapStrFloat => LL::IterBeginStrFloat(dst_reg.into(), arr_reg.into()),
+                    Ty::MapStrStr => LL::IterBeginStrStr(dst_reg.into(), arr_reg.into()),
                     Ty::Int | Ty::Float | Ty::Str | Ty::IterInt | Ty::IterStr => {
                         // covered by the error check above
                         unreachable!()
@@ -972,12 +1013,12 @@ impl<'a, 'b> View<'a, 'b> {
                 let target_reg = if dst_ty == Ty::Int {
                     dst_reg
                 } else {
-                    reg_of_ty!(self.regs, Ty::Int)
+                    self.regs.stats.reg_of_ty(Ty::Int)
                 };
                 let (iter_reg, iter_ty) = self.get_reg(pv)?;
-                self.push(match iter_ty.iter()? {
-                    Ty::Int => Instr::IterHasNextInt(target_reg.into(), iter_reg.into()),
-                    Ty::Str => Instr::IterHasNextStr(target_reg.into(), iter_reg.into()),
+                self.pushl(match iter_ty.iter()? {
+                    Ty::Int => LL::IterHasNextInt(target_reg.into(), iter_reg.into()),
+                    Ty::Str => LL::IterHasNextStr(target_reg.into(), iter_reg.into()),
                     _ => unreachable!(),
                 });
                 self.convert(dst_reg, dst_ty, target_reg, Ty::Int)?
@@ -988,11 +1029,11 @@ impl<'a, 'b> View<'a, 'b> {
                 let target_reg = if dst_ty == elt_ty {
                     dst_reg
                 } else {
-                    reg_of_ty!(self.regs, elt_ty)
+                    self.regs.stats.reg_of_ty(elt_ty)
                 };
-                self.push(match elt_ty {
-                    Ty::Int => Instr::IterGetNextInt(target_reg.into(), iter_reg.into()),
-                    Ty::Str => Instr::IterGetNextStr(target_reg.into(), iter_reg.into()),
+                self.pushl(match elt_ty {
+                    Ty::Int => LL::IterGetNextInt(target_reg.into(), iter_reg.into()),
+                    Ty::Str => LL::IterGetNextStr(target_reg.into(), iter_reg.into()),
                     _ => unreachable!(),
                 });
                 self.convert(dst_reg, dst_ty, target_reg, elt_ty)?
@@ -1002,12 +1043,12 @@ impl<'a, 'b> View<'a, 'b> {
                 let target_reg = if target_ty == dst_ty {
                     dst_reg
                 } else {
-                    reg_of_ty!(self.regs, target_ty)
+                    self.regs.stats.reg_of_ty(target_ty)
                 };
-                self.push(match target_ty {
-                    Ty::Str => Instr::LoadVarStr(target_reg.into(), *bv),
-                    Ty::Int => Instr::LoadVarInt(target_reg.into(), *bv),
-                    Ty::MapIntStr => Instr::LoadVarIntMap(target_reg.into(), *bv),
+                self.pushl(match target_ty {
+                    Ty::Str => LL::LoadVarStr(target_reg.into(), *bv),
+                    Ty::Int => LL::LoadVarInt(target_reg.into(), *bv),
+                    Ty::MapIntStr => LL::LoadVarIntMap(target_reg.into(), *bv),
                     _ => unreachable!(),
                 });
                 self.convert(dst_reg, dst_ty, target_reg, target_ty)?
@@ -1015,6 +1056,7 @@ impl<'a, 'b> View<'a, 'b> {
         };
         Ok(())
     }
+
     fn stmt(&mut self, stmt: &cfg::PrimStmt<'a>) -> Result<()> {
         match stmt {
             PrimStmt::AsgnIndex(arr, pv, pe) => {
@@ -1022,21 +1064,21 @@ impl<'a, 'b> View<'a, 'b> {
                 let (mut k_reg, k_ty) = self.get_reg(pv)?;
                 let a_key_ty = a_ty.key()?;
                 if k_ty != a_key_ty {
-                    let conv_reg = reg_of_ty!(self.regs, a_key_ty);
+                    let conv_reg = self.regs.stats.reg_of_ty(a_key_ty);
                     self.convert(conv_reg, a_key_ty, k_reg, k_ty)?;
                     k_reg = conv_reg;
                 }
                 let v_ty = a_ty.val()?;
-                let v_reg = reg_of_ty!(self.regs, v_ty);
+                let v_reg = self.regs.stats.reg_of_ty(v_ty);
                 self.expr(v_reg, v_ty, pe)?;
                 use Ty::*;
-                self.push(match a_ty {
-                    MapIntInt => Instr::StoreIntInt(a_reg.into(), k_reg.into(), v_reg.into()),
-                    MapIntFloat => Instr::StoreIntFloat(a_reg.into(), k_reg.into(), v_reg.into()),
-                    MapIntStr => Instr::StoreIntStr(a_reg.into(), k_reg.into(), v_reg.into()),
-                    MapStrInt => Instr::StoreStrInt(a_reg.into(), k_reg.into(), v_reg.into()),
-                    MapStrFloat => Instr::StoreStrFloat(a_reg.into(), k_reg.into(), v_reg.into()),
-                    MapStrStr => Instr::StoreStrStr(a_reg.into(), k_reg.into(), v_reg.into()),
+                self.pushl(match a_ty {
+                    MapIntInt => LL::StoreIntInt(a_reg.into(), k_reg.into(), v_reg.into()),
+                    MapIntFloat => LL::StoreIntFloat(a_reg.into(), k_reg.into(), v_reg.into()),
+                    MapIntStr => LL::StoreIntStr(a_reg.into(), k_reg.into(), v_reg.into()),
+                    MapStrInt => LL::StoreStrInt(a_reg.into(), k_reg.into(), v_reg.into()),
+                    MapStrFloat => LL::StoreStrFloat(a_reg.into(), k_reg.into(), v_reg.into()),
+                    MapStrStr => LL::StoreStrStr(a_reg.into(), k_reg.into(), v_reg.into()),
                     Int | Float | Str | IterInt | IterStr => {
                         return err!(
                             "in stmt {:?} computed type is non-map type {:?}",
@@ -1052,45 +1094,28 @@ impl<'a, 'b> View<'a, 'b> {
             }
             PrimStmt::SetBuiltin(v, pe) => {
                 let ty = Ty::of_var(*v);
-                let reg = reg_of_ty!(self.regs, ty);
+                let reg = self.regs.stats.reg_of_ty(ty);
                 self.expr(reg, ty, pe)?;
                 use Ty::*;
-                self.push(match ty {
-                    Str => Instr::StoreVarStr(*v, reg.into()),
-                    MapIntStr => Instr::StoreVarIntMap(*v, reg.into()),
-                    Int => Instr::StoreVarInt(*v, reg.into()),
+                self.pushl(match ty {
+                    Str => LL::StoreVarStr(*v, reg.into()),
+                    MapIntStr => LL::StoreVarIntMap(*v, reg.into()),
+                    Int => LL::StoreVarInt(*v, reg.into()),
                     _ => return err!("unexpected type for variable {} : {:?}", v, ty),
                 });
             }
             PrimStmt::Return(v) => {
                 let (v_reg, v_ty) = self.get_reg(v)?;
-                let FuncInfo {
-                    ret_reg, ret_ty, ..
-                } = &self.func_info[self.frame.cur_ident as usize];
-                self.convert(*ret_reg, *ret_ty, v_reg, v_ty)?;
-                let ret_index = self.push_with_len(Instr::Ret);
-                self.frame.rets.insert(ret_index, Some((*ret_reg, *ret_ty)));
+                let ret_ty = self.func_info[self.frame.cur_ident as usize].ret_ty;
+                if v_ty != ret_ty {
+                    let ret_reg = self.regs.stats.reg_of_ty(ret_ty);
+                    self.convert(ret_reg, ret_ty, v_reg, v_ty)?;
+                    self.pushr(HighLevel::Ret(ret_reg, ret_ty));
+                } else {
+                    self.pushr(HighLevel::Ret(v_reg, v_ty));
+                }
             }
         };
         Ok(())
     }
-
-    fn push(&mut self, i: Instr<'a>) {
-        self.frame.instrs.push(i)
-    }
-    fn push_with_len(&mut self, i: Instr<'a>) -> usize {
-        let res = self.frame.instrs.len();
-        self.push(i);
-        res
-    }
-}
-
-pub(crate) fn bytecode<'a>(
-    ctx: &cfg::ProgramContext<'a, &'a str>,
-    // default to std::io::stdin()
-    reader: impl std::io::Read + 'static,
-    // default to std::io::BufWriter::new(std::io::stdout())
-    writer: impl std::io::Write + 'static,
-) -> Result<Interp<'a>> {
-    Ok(ProgramGenerator::init_from_ctx(ctx)?.into_interp(reader, writer))
 }
